impl blocking read
Running strace on Kiro reveals that it polls every 100ms for an input
returning 0 if nothing was read. This pull request implements a blocking
read for noncanonical mode as specified in the canonical and
noncanonical section in the termios man page:
https://man7.org/linux/man-pages/man3/termios.3.html

This provides a more efficient reading of the input, by blocking after
an empty read of 0 bytes without wasteful polling cycles.